Move functions to Functions folder and namespace
