+++ FastPass/FastPass.API/Functions/TextAnalyticsServiceProxyFunction.cs
@@ -6,7 +6,7 @@ using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using System.Net;
 
-namespace FastPass.Api;
+namespace FastPass.API.Functions;
 
 public class TextAnalyticsServiceProxyFunction
 {
